+++ examples/tls/client-tls.c
@@ -143,7 +143,7 @@ int main(int argc, char** argv)
     len = strnlen(buff, sizeof(buff));
 
     /* Send the message to the server */
-    if ((size_t) (ret = wolfSSL_write(ssl, buff, len)) != len) {
+    if ((size_t) (ret = wolfSSL_write(ssl, buff, (int) len)) != len) {
         fprintf(stderr, "ERROR: failed to write entire message\n");
         fprintf(stderr, "%d bytes of %d bytes were sent", ret, (int) len);
         goto cleanup;
